adding documentation and organization
