+++ app/src/main/java/com/example/clicker/network/domain/BetterTTVEmotes.kt
@@ -1,8 +1,8 @@
 package com.example.clicker.network.domain
 
-import com.example.clicker.network.clients.IndivBetterTTVEmote
-import com.example.clicker.network.models.twitchAuthentication.ValidatedUser
-import com.example.clicker.util.NetworkNewUserResponse
+
+import com.example.clicker.network.models.emotes.IndivBetterTTVEmote
+
 import com.example.clicker.util.Response
 import kotlinx.coroutines.flow.Flow
