+++ app/src/main/java/com/example/clicker/network/repository/TwitchEmotImpl.kt
@@ -20,15 +20,23 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.example.clicker.R
-import com.example.clicker.network.clients.BetterTTVChannelEmote
-import com.example.clicker.network.clients.BetterTTVChannelEmotes
+
 import com.example.clicker.network.clients.BetterTTVEmoteClient
-import com.example.clicker.network.clients.BetterTTVSharedEmote
 import com.example.clicker.network.clients.ChannelEmote
-import com.example.clicker.network.clients.IndivBetterTTVEmote
 import com.example.clicker.network.clients.TwitchEmoteClient
 import com.example.clicker.network.domain.BetterTTVEmotes
 import com.example.clicker.network.domain.TwitchEmoteRepo
+import com.example.clicker.network.models.emotes.BetterTTVIndivChannelEmote
+import com.example.clicker.network.models.emotes.BetterTTVChannelEmotes
+import com.example.clicker.network.models.emotes.BetterTTVSharedEmote
+import com.example.clicker.network.models.emotes.IndivBetterTTVEmote
+import com.example.clicker.network.repository.models.EmoteListMap
+import com.example.clicker.network.repository.models.EmoteNameUrl
+import com.example.clicker.network.repository.models.EmoteNameUrlEmoteType
+import com.example.clicker.network.repository.models.EmoteNameUrlEmoteTypeList
+import com.example.clicker.network.repository.models.EmoteNameUrlList
+import com.example.clicker.network.repository.models.EmoteTypes
+import com.example.clicker.network.repository.models.IndivBetterTTVEmoteList
 import com.example.clicker.network.repository.util.EmoteParsing
 import com.example.clicker.network.repository.util.handleException
 import com.example.clicker.presentation.stream.views.chat.chatSettings.ChatBadgePair
@@ -425,7 +433,7 @@ class TwitchEmoteImpl @Inject constructor(
      * emitIndivBetterTTVChannelEmotes parses a list of [BetterTTVChannelEmote] objects and emits a
      * list shown to the user in the emote board
      * */
-    private fun emitIndivBetterTTVChannelEmotes(listOfChannelEmotes: List<BetterTTVChannelEmote>){
+    private fun emitIndivBetterTTVChannelEmotes(listOfChannelEmotes: List<BetterTTVIndivChannelEmote>){
         val parsedChannelEmotes =listOfChannelEmotes.map {channelEmote ->
             IndivBetterTTVEmote(
                 id =channelEmote.id,
@@ -501,79 +509,3 @@ class TwitchEmoteImpl @Inject constructor(
 }
 
 
-/**
- * EmoteNameUrl represents a single Twitch Emote from the Twitch servers. Each instance of this class is a unique Emote
- *
- * @param name the name of the Twitch emote
- * @param url the url that is hosted on the twitch servers and is what we use to load the image
- * */
-data class EmoteNameUrl(
-    val name:String,
-    val url:String,
-)
-
-/**
- * EmoteNameEmoteType represents a single Twitch Emote from the Twitch servers, when calling get channel emotes
- * - you can read more about getting channel emotes, [HERE](https://dev.twitch.tv/docs/api/reference/#get-channel-emotes)
- *
- * @param name the name of the Twitch emote
- * @param url the url that is hosted on the twitch servers and is what we use to load the image,
- * @param emoteType a [EmoteTypes] used to represent the type of emote that it is
- * */
-data class EmoteNameUrlEmoteType(
-    val name:String,
-    val url:String,
-    val emoteType:EmoteTypes
-)
-
-/**
- * EmoteTypes represents the two types of emotes, subscribers and followers
- * */
-enum class EmoteTypes {
-    SUBS, FOLLOWERS,
-}
-
-@Immutable
-data class EmoteNameUrlList(
- val list:List<EmoteNameUrl> = listOf()
-)
-
-@Immutable
-data class EmoteListMap(
-    val map:Map<String, InlineTextContent>
-)
-
-@Immutable
-data class EmoteNameUrlEmoteTypeList(
-    val list:List<EmoteNameUrlEmoteType> = listOf()
-)
-
-/**
- * EmoteNameUrlNumberList
- * */
-@Immutable
-data class EmoteNameUrlNumberList(
-    val list:List<EmoteNameUrlNumber> = listOf()
-)
-
-/**
- * class to show the list of individual channel emotes of BetterTTV
- * */
-@Immutable
-data class IndivBetterTTVEmoteList(
-    val list: List<IndivBetterTTVEmote> = listOf()
-)
-
-/**
- * EmoteNameUrlNumber represents a single Twitch Emote from the Twitch servers and the number of times it was clicked.
- * This data class is used soley for the purpose of the most frequently clicked emotes
- *
- * @param name the name of the Twitch emote
- * @param url the url that is hosted on the twitch servers and is what we use to load the image
- * @param timesClicked the number of times this emote was clicked inside of the Twitch emote board
- * */
-data class EmoteNameUrlNumber(
-    val name:String,
-    val url:String,
-    val timesClicked:Int
-)
