+++ app/src/main/java/com/example/clicker/presentation/stream/views/chat/FullChatModView.kt
@@ -11,15 +11,13 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
-import com.example.clicker.network.clients.Emote
-import com.example.clicker.network.clients.IndivBetterTTVEmote
 import com.example.clicker.network.models.websockets.TwitchUserData
-import com.example.clicker.network.repository.EmoteListMap
-import com.example.clicker.network.repository.EmoteNameUrl
-import com.example.clicker.network.repository.EmoteNameUrlEmoteTypeList
-import com.example.clicker.network.repository.EmoteNameUrlList
-import com.example.clicker.network.repository.EmoteNameUrlNumberList
-import com.example.clicker.network.repository.IndivBetterTTVEmoteList
+import com.example.clicker.network.repository.models.EmoteListMap
+import com.example.clicker.network.repository.models.EmoteNameUrl
+import com.example.clicker.network.repository.models.EmoteNameUrlEmoteTypeList
+import com.example.clicker.network.repository.models.EmoteNameUrlList
+import com.example.clicker.network.repository.models.IndivBetterTTVEmoteList
+
 import com.example.clicker.presentation.stream.util.FilteredChatListImmutableCollection
 import com.example.clicker.presentation.stream.util.ForwardSlashCommands
 import com.example.clicker.presentation.stream.util.ForwardSlashCommandsImmutableCollection
@@ -50,7 +48,7 @@ fun FullChatModView(
     noChat:Boolean,
     deleteChatMessage:(String)->Unit,
     clickedCommandAutoCompleteText: (String) -> Unit,
-    badgeListMap:EmoteListMap,
+    badgeListMap: EmoteListMap,
     hideSoftKeyboard:()-> Unit,
     emoteBoardGlobalList: EmoteNameUrlList,
     emoteBoardChannelList: EmoteNameUrlEmoteTypeList,
@@ -80,8 +78,7 @@ fun FullChatModView(
     sharedBetterTTVEmoteContentMap: EmoteListMap,
     lowPowerMode:Boolean,
 
-
-){
+    ){
     val lazyColumnListState = rememberLazyListState()
     var autoscroll by remember { mutableStateOf(true) }
     val emoteKeyBoardHeight = remember { mutableStateOf(0.dp) }
