+++ app/src/test/java/com/example/clicker/network/repository/util/EmoteParsingTest.kt
@@ -1,27 +1,10 @@
 package com.example.clicker.network.repository.util
 
 import androidx.compose.foundation.text.InlineTextContent
-import com.example.clicker.network.repository.EmoteNameUrl
 import com.example.clicker.presentation.modView.BoxTypeIndex
 import org.junit.Assert
 import org.junit.Test
 
 class EmoteParsingTest {
-    val underTest = EmoteParsing()
-
-
-    @Test
-    fun `emote parsing adds emote value to map`(){
-        /**GIVEN*/
-        val nameUrl = EmoteNameUrl("testing","testingURL")
-        val innerInlineContentMap: MutableMap<String, InlineTextContent> = mutableMapOf()
-
-        /**WHEN*/
-        underTest.createMapValueForComposeChat(nameUrl,innerInlineContentMap)
-
-        /**THEN*/
-
-        Assert.assertEquals(1, innerInlineContentMap.size)
-    }
 
 }
